+++ org.eclipse.winery.repository.ui/src/app/wineryMainModules/artifactTypes/artifactType.module.ts
@@ -12,11 +12,15 @@
 import { NgModule } from '@angular/core';
 import { ArtifactTypeRouterModule } from './artifactTypeRouter.module';
 import { CommonModule } from '@angular/common';
+import { WineryLicenseModule } from '../../wineryLicenseModule/wineryLicense.module';
+import { WineryReadmeModule } from '../../wineryReadmeModule/wineryReadme.module';
 
 @NgModule({
     imports: [
         CommonModule,
         ArtifactTypeRouterModule,
+        WineryReadmeModule,
+        WineryLicenseModule
     ],
 })
 export class ArtifactTypeModule {